Create the Authentication middleware
